Fix CLI --coops, --cobis options when files not in CWD
The filename detection currently relies on an exact string match,
leading to errors if --cohpcar was not in the current directory.

Here we use Pathlib to interpret the path, compare only the filename
and set an appropriate parent directory when searching for other
files. Pymatgen doesn't seem to mind being passed a Path directly; it
might be worth using these objects more systematically.