Rework argparse setup: plot/automaticplot/description subcommands
Some more argparse features are used to express the relationships
between different arguments. This moves exclusive-argument and
available-argument logic from main() to the parser setup, and should
give more helpful --help pages.

- The subcommands feature allows the three different run-modes to be
  expressed more clearly. Each has its own help page, with only
  relevant arguments.
- "Parent" parsers are mixed to avoid redundancy in setting this up
- Argument groups bring together related arguments, making --help
  easier to navigate.

The downside is that it may not be 100% obvious that users need to
type e.g. `lobsterpy plot -h` to see plot options. I have tried to
explain this in the `lobsterpy -h` page...

This is not completely compatible with the previous syntax, but the
changes are small and motivated:

--plot, --automaticplot, --description keyword arguments have been
replaced by a positional "action" argument which should be set as
"plot", "automatic-plot", or "description".

--plot previously required values to be assigned to it. Now there is a
positional argument when the "plot" action is chosen. So now instead
of e.g.

  lobsterpy --poscar /path/to/poscar --plot 1 2 --coop ...

we must have

  lobsterpy plot 1 2 --poscar /path/to/poscar --coop ...

While this arguably is a slight burden on the user, hopefully it will
pay off with easier-to-read command samples. (As we can now always
check the second word to see what mode we are in, which is important
context for other arguments.)