CLI Tests: check plot content
A few bits of plot information are checked as well as the line values,
so we know that styling did something.

It is necessary to explicitly reset the styling, because the
command-line tools use matplotlib.style.use which sets the style going
forwards. This problem could be avoided with matplotlib.syle.context,
but

- it doesn't matter for normal CLI use because the python session will
  end after plotting

- the style contexts don't work properly for setting fonts. (Issue
  here, marked WONTFIX:
  https://github.com/matplotlib/matplotlib/issues/11673)