Provide simple Gaussian broadening.
The implementation uses convolution on a regular energy grid; this is
more efficient than evaluating the Guassian at every point, but
invalid for irregular energy grids. I have included a check for this
condition, but due to the limited precision of input file data there
is quite a generous tolerance of 1e-5.